Add functionality to use a fact to specify os_user for oracle and asm. Fixes #36
Adding a fact ORA_OS_USER overrides the default account used for connecting to Oracle.

Adding the fact ASM_OS_USER overrides the default account used for connecting to the ASM instance.

For masterless puppet, you can also use a symbol with the same name. The symbol has precedence over the fact.

Check [this article](https://docs.puppetlabs.com/facter/2.3/custom_facts.html) on how to define a custom fact for a system.

BEWARE: This functionality will probably be overriden by a news defaults system in the future. For now it is just a quick fix